# Conflicts:
#	lib/Backend/Models/Absence.dart
#	lib/Backend/Models/LoginModel.dart

+++ lib/Backend/Models/UserModel.dart
@@ -10,7 +10,7 @@ class UserModel {
   final String? departmentDescription;
   final String? jobCode;
   final String? jobDescription;
-  final double? leaveBalance;
+  final int? leaveBalance;
   final List<Role>? roles;
   final String? selectedRoleId;
   final String token;
